% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnbfxrate.R
\name{cnbfxrate}
\alias{cnbfxrate}
\title{Central bank exchange rate fixing}
\usage{
cnbfxrate(date = Sys.Date() - 1, currency_code = "ALL")
}
\arguments{
\item{date}{Date of publication as date, default is yesterday.}

\item{currency_code}{ISO Code of currency, default is ALL / complete list.}
}
\value{
data frame - date_valid, country, currency name, quoted amount, currency ISO code, rate
}
\description{
A function returning data frame of FX rates as fixed by the central bank.
}
\examples{
cnbfxrate(as.Date("2002-08-12"))



}
