% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czechrates.R
\docType{package}
\name{czechrates-package}
\alias{czechrates-package}
\title{czechrates: Czech Interest Rates}
\description{
Interface to interest rates as published by the Czech National Bank.
}
\keyword{internal}
