% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{downloader}
\alias{downloader}
\title{Internal function - generic downloader of raw data from www.cnb.cz}
\usage{
downloader(year, verbose = FALSE)
}
\arguments{
\item{year}{file to be downloaded}

\item{verbose}{should progress messages be displayed? Default is not.}
}
\value{
data frame of annual pribor & pribid rates
}
\description{
Internal function - generic downloader of raw data from www.cnb.cz
}
\keyword{internal}
