% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pribor.R
\name{pribor}
\alias{pribor}
\title{pribor}
\arguments{
\item{date}{Date of fixing as date, default is yesterday.}

\item{maturity}{Maturity of loan as string, default is overnight ("1D").}
}
\value{
data frame - first column is date, second is relevant PRIBOR rate.
}
\description{
A function returning data frame of PRIBOR rates.
}
\details{
The function expects date input, and returns data frame of two columns - date, and relevant PRIBOR rate (as determined by `maturity` parameter).

PRIBOR rates are reported as fractions, i.e. not as percentages (1% = .01).

For dates when no PRIBOR was quoted (e.g. Bank Holidays, such as December 24th on any year, or August 13th, 2002 when no PRIBOR was quoted due to catastrophic floods) no result will be returned.
}
\examples{
pribor(as.Date("2002-08-12"), "1D")



}
