% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnbfxrate.R
\name{cnbfxrate}
\alias{cnbfxrate}
\title{Central bank exchange rate fixing}
\usage{
cnbfxrate(date = Sys.Date() - 1, specific_currency = "ALL")
}
\arguments{
\item{date}{Date of publication as date, default is yesterday.}

\item{specific_currency}{ISO Code of currency, default is ALL / complete list.}
}
\value{
a data frame - date_valid, currency ISO code, quoted amount, rate
}
\description{
A function returning data frame of FX rates as fixed by the central bank.
}
\examples{
cnbfxrate(as.Date("2002-08-12"), "EUR") # EUR/CZK rate for August 12th, 2002



}
